reverting patches for zmq-57 (causes problem on Linux/x86_64)
git-svn-id: http://svn.fastmq.org/svn/zeromq/trunk@618 7451a71c-b6c8-49cc-8254-141fc5361860

+++ jzmq/Jzmq.cpp
@@ -63,19 +63,17 @@ JNIEXPORT void JNICALL Java_Jzmq_construct (JNIEnv *env, jobject obj,
     env->ReleaseStringUTFChars (hostname_, hostname);
 
     //  Set the context.
-    env->SetLongField (obj, context_fid, (jlong)(unsigned int) context);
-
-
-	}
+    env->SetLongField (obj, context_fid, (jlong) context);
+}
 
 JNIEXPORT void JNICALL Java_Jzmq_finalize (JNIEnv *env, jobject obj)
-	{
-	    //  Get the context.
-	    context_t *context = (context_t*)(unsigned int) env->GetLongField (obj, context_fid);
-	    assert (context);
+{
+    //  Get the context.
+    context_t *context = (context_t*) env->GetLongField (obj, context_fid);
+    assert (context);
 
-	    //  Deallocate the 0MQ infrastructure.
-	    delete context->dispatcher;
+    //  Deallocate the 0MQ infrastructure.
+    delete context->dispatcher;
     delete context->locator;
     delete context;
 }
@@ -84,7 +82,7 @@ JNIEXPORT jint JNICALL Java_Jzmq_createExchange (JNIEnv *env, jobject obj,
     jstring exchange_, jint scope_, jstring nic_)
 {
     //  Get the context.
-    context_t *context = (context_t*)(unsigned int) env->GetLongField (obj, context_fid);
+    context_t *context = (context_t*) env->GetLongField (obj, context_fid);
     assert (context);
 
     //  Get the exchange name.
@@ -117,7 +115,7 @@ JNIEXPORT jint JNICALL Java_Jzmq_createQueue (JNIEnv *env, jobject obj,
     jstring queue_, jint scope_, jstring nic_)
 {
     //  Get the context.
-    context_t *context = (context_t*)(unsigned int) env->GetLongField (obj, context_fid);
+    context_t *context = (context_t*) env->GetLongField (obj, context_fid);
     assert (context);
 
     //  Get the queue name.
@@ -150,7 +148,7 @@ JNIEXPORT void JNICALL Java_Jzmq_bind (JNIEnv *env, jobject obj,
     jstring exchange_, jstring queue_)
 {
     //  Get the context.
-    context_t *context = (context_t*)(unsigned int) env->GetLongField (obj, context_fid);
+    context_t *context = (context_t*) env->GetLongField (obj, context_fid);
     assert (context);
 
     //  Get the exchange name.
@@ -175,7 +173,7 @@ JNIEXPORT void JNICALL Java_Jzmq_send (JNIEnv *env, jobject obj,
     jint eid_, jbyteArray data_)
 {
     //  Get the context.
-    context_t *context = (context_t*)(unsigned int) env->GetLongField (obj, context_fid);
+    context_t *context = (context_t*) env->GetLongField (obj, context_fid);
     assert (context);
 
     //  Get the data from the bytearray.
@@ -193,7 +191,7 @@ JNIEXPORT void JNICALL Java_Jzmq_send (JNIEnv *env, jobject obj,
 JNIEXPORT jbyteArray JNICALL Java_Jzmq_receive (JNIEnv *env, jobject obj)
 {
     //  Get the context.
-    context_t *context = (context_t*)(unsigned int) env->GetLongField (obj, context_fid);
+    context_t *context = (context_t*) env->GetLongField (obj, context_fid);
     assert (context);
 
     //  Get new message.
